feat : favicon 추가 및 비회원 처리
